pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fairfield_housing_analysis"
version = "0.1.0"
description = "Fairfield County Housing Market Analysis Application"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.31.0",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "plotly>=5.18.0",
    "matplotlib>=3.8.2",
    "altair>=5.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.2",
    "statsmodels>=0.14.0",
    "prophet>=1.1.4",
    "openai>=1.3.5",
    "anthropic>=0.8.1",
    "fpdf>=1.7.2",
    "python-docx>=1.0.1",
    "streamlit-extras>=0.3.5",
    "streamlit-shadcn-ui>=0.1.7",
    "st-pages>=0.4.5",
    "pyyaml>=6.0.1",
    "tabulate>=0.9.0",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
